+++ src/gfxlib2/xbox/gfx_driver.c
@@ -159,7 +159,7 @@ const GFXDRIVER *__fb_gfx_drivers_list[] = {
 };
 
 /*:::::*/
-void fb_hScreenInfo(int *width, int *height, int *depth, int *refresh)
+void fb_hScreenInfo(ssize_t *width, ssize_t *height, ssize_t *depth, ssize_t *refresh)
 {
 	/* !!!FIXME!!! */
 	VIDEO_MODE vm;
